Move WorkflowContentType to workflows submodule
